smart-crm/requirements.txt
# Flask and Extensions
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
python-dotenv>=1.0.0
pydantic>=2.5.0

# Authentication
bcrypt>=4.1.0
PyJWT>=2.8.0

# Database
supabase>=2.0.0

# AI/ML - CrewAI
crewai>=0.80.0
crewai-tools>=0.14.0
openai>=1.50.0
langchain-openai>=0.1.0

# Vector Database
chromadb>=0.4.0

# Utilities
python-dateutil>=2.8.0
uuid6>=2024.1.0
